51703 : Onboarding screen displayed for few second before the display of card screen (#81)
