+++ app/src/main/res/drawable/splash_background.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:top="0dp"
+        android:right="0dp"
+        android:left="0dp"
+        android:bottom="0dp"
+        android:drawable="@color/eXoWhiteColor">
+    </item>
+    <item
+        android:drawable="@mipmap/ic_launcher"
+        android:height="60dp"
+        android:width="60dp"
+        android:gravity="center">
+    </item>
+</layer-list>
